requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.0
PyJWT==2.9.0
python-multipart==0.0.9
jinja2==3.1.4
passlib==1.7.4
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
torch>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
tqdm>=4.65.0
